+++ todo/src/components/ToDoList.tsx
@@ -1,0 +1,58 @@
+import { useForm } from "react-hook-form";
+import {
+  atom,
+  useRecoilState,
+  useRecoilValue,
+  useSetRecoilState,
+} from "recoil";
+
+interface IForm {
+  toDo: string;
+}
+
+interface IToDo {
+  text: string;
+  category: "DONE" | "DOING" | "TO_DO";
+  id: number;
+}
+
+const toDoState = atom<IToDo[]>({
+  key: "toDo",
+
+  default: [],
+});
+
+function ToDoList() {
+  const [toDos, setToDos] = useRecoilState(toDoState);
+  const { register, handleSubmit, setValue } = useForm<IForm>();
+  const handleValid = ({ toDo }: IForm) => {
+    setToDos((oldToDos) => [
+      { text: toDo, category: "TO_DO", id: Date.now() },
+      ...oldToDos,
+    ]);
+    setValue("toDo", "");
+  };
+  console.log(toDos);
+  return (
+    <div>
+      <h1>To Dos</h1>
+      <hr />
+      <form onSubmit={handleSubmit(handleValid)}>
+        <input
+          {...register("toDo", {
+            required: "Please write a To Do",
+          })}
+          placeholder="Write a to do"
+        />
+        <button>Add</button>
+      </form>
+      <ul>
+        {toDos.map((toDo) => (
+          <li key={toDo.id}>{toDo.text}</li>
+        ))}
+      </ul>
+    </div>
+  );
+}
+
+export default ToDoList();
